First commit, with base features, sign up login, etc.

+++ src/main/java/Service/cart/CartService.java
@@ -1,11 +1,12 @@
 package Service.cart;
 
 import Entity.Cart;
+
 import java.util.List;
 
 public interface CartService {
-    Cart addToCart(int userId, int productId, int quantity); // Add or update a cart item
-    List<Cart> getAllCartItemsForUser(int userId); // Get all cart items for a user
-    void removeCartItem(int userId, int productId); // Remove specific cart item
-    void clearCartForUser(int userId); // Clear all cart items for a user
+    Cart addToCart(Long userId, Long productId, int quantity);
+    List<Cart> getAllCartItemsForUser(Long userId);
+    void removeCartItem(Long userId, Long productId);
+    void clearCartForUser(Long userId);
 }
